refactor :  패키지 구조 변경, 싱글톤 적용, view 유틸클래스로 변경
- 패키지 구조에 service 생성
- LottoMachine 을 싱글톤으로 변경
- InputView 를 유틸클래스로 변경
