+++ src/test/java/domain/LottoMachineTest.java
@@ -4,11 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
+import service.LottoMachine;
+
 public class LottoMachineTest {
 
 	@Test
 	void 로또_발행(){
-		assertThat(new LottoMachine()
+		assertThat(LottoMachine.getInstance()
 			.createLottos(2))
 			.isInstanceOf(Lottos.class);
 	}
